+++ esnext/lib/history-util.js
@@ -6,14 +6,26 @@ const historyUtil = {
 	// Get History Section
 	getHistorySection: function (opts) {
 		// Prepare
-		const file = 'HISTORY.md'
-		const url = projectzUtil.getFileUrl(opts, file)
+		let link = null
+		const file = opts.filenamesForReadmeFiles.history
+		if ( file ) {
+			link = projectzUtil.getLink({
+				url: projectzUtil.getFileUrl(opts, file),
+				text: `Discover the release history by heading on over to the <code>${file}</code> file.`
+			})
+		}
+		else if ( opts.github.slug ) {
+			link = projectzUtil.getLink({
+				url: `https://github.com/${opts.github.slug}/releases`,
+				text: `Discover the release history by heading on over to the releases page.`
+			})
+		}
+		else {
+			throw new Error('History section either requires a HISTORY file or a Github repository')
+		}
 
 		// Return
-		return [
-			'## History',
-			`[Discover the change history by heading on over to the \`${file}\` file.](${url})`
-		].join('\n')
+		return '<h2>History</h2>\n\n' + link
 	}
 }